requirements.txt
# Bot Framework
botbuilder-core==4.14.3
botbuilder-schema==4.14.3
flask==2.2.5

# Microsoft Graph API 연동
requests==2.31.0
msal==1.26.0  # OAuth2를 사용할 경우 필요

# Azure Blob Storage
azure-storage-blob>=12.19.0
azure-core==1.30.1

# OpenAI API (요약 기능)
openai==1.30.1

# 기타
python-dotenv==1.0.1
beautifulsoup4>=4.12.0
lxml>=4.9.0

# FastAPI 서버 (병렬 또는 이후 전환용)
fastapi==0.110.0
uvicorn==0.29.0